Enable HTTP trailer support
Swoole supports trailer response headers, but Laravel does not (same as almost all other PHP frameworks). This presents a challenge, where we either need to extend Laravel's (Symfony's to be precise) HTTP response object to include trailers support or take a different strategy. Here, the proposal is to take a different approach, i.e. rely on the `trailer` header to specify the list of all the headers that should be converted into trailing. Similar approach has been taken for RoadRunner application server and worked well, so the hope is that this is a reasonable solution that is backwards compatible with Laravel's response and supports a new feature.